fix: move StateProvider read functions to AccountReader
