+++ crates/storage/storage-api/src/account.rs
@@ -1,21 +1,74 @@
+use crate::BytecodeReader;
 use alloc::{
     collections::{BTreeMap, BTreeSet},
     vec::Vec,
 };
-use alloy_primitives::{Address, BlockNumber};
+use alloy_consensus::constants::KECCAK_EMPTY;
+use alloy_primitives::{Address, BlockNumber, StorageKey, StorageValue, U256};
 use auto_impl::auto_impl;
 use core::ops::{RangeBounds, RangeInclusive};
 use reth_db_models::AccountBeforeTx;
-use reth_primitives_traits::Account;
+use reth_primitives_traits::{Account, Bytecode};
 use reth_storage_errors::provider::ProviderResult;
 
 /// Account reader
 #[auto_impl(&, Arc, Box)]
-pub trait AccountReader {
+pub trait AccountReader: BytecodeReader {
     /// Get basic account information.
     ///
     /// Returns `None` if the account doesn't exist.
     fn basic_account(&self, address: &Address) -> ProviderResult<Option<Account>>;
+
+    /// Get storage of given account.
+    ///
+    /// Returns `None` if the account doesn't exist or the storage key is not found.
+    fn storage(
+        &self,
+        account: Address,
+        storage_key: StorageKey,
+    ) -> ProviderResult<Option<StorageValue>>;
+
+    /// Get account code by its address.
+    ///
+    /// Returns `None` if the account doesn't exist or account is not a contract
+    fn account_code(&self, addr: &Address) -> ProviderResult<Option<Bytecode>> {
+        // Get basic account information
+        // Returns None if acc doesn't exist
+        let acc = match self.basic_account(addr)? {
+            Some(acc) => acc,
+            None => return Ok(None),
+        };
+
+        if let Some(code_hash) = acc.bytecode_hash {
+            if code_hash == KECCAK_EMPTY {
+                return Ok(None)
+            }
+            // Get the code from the code hash
+            return self.bytecode_by_hash(&code_hash)
+        }
+
+        // Return `None` if no code hash is set
+        Ok(None)
+    }
+
+    /// Get account balance by its address.
+    ///
+    /// Returns `None` if the account doesn't exist
+    fn account_balance(&self, addr: &Address) -> ProviderResult<Option<U256>> {
+        // Get basic account information
+        // Returns None if acc doesn't exist
+
+        self.basic_account(addr)?.map_or_else(|| Ok(None), |acc| Ok(Some(acc.balance)))
+    }
+
+    /// Get account nonce by its address.
+    ///
+    /// Returns `None` if the account doesn't exist
+    fn account_nonce(&self, addr: &Address) -> ProviderResult<Option<u64>> {
+        // Get basic account information
+        // Returns None if acc doesn't exist
+        self.basic_account(addr)?.map_or_else(|| Ok(None), |acc| Ok(Some(acc.nonce)))
+    }
 }
 
 /// Account reader
